move validateStorage
