+++ src/funcConfig/local.settings.ts
@@ -3,14 +3,10 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { AzExtFsExtra, AzureWizard, DialogResponses, IActionContext, parseError } from '@microsoft/vscode-azext-utils';
+import { AzExtFsExtra, DialogResponses, IActionContext, parseError } from '@microsoft/vscode-azext-utils';
 import * as path from 'path';
 import * as vscode from 'vscode';
-import { AzureWebJobsStorageExecuteStep } from '../commands/appSettings/connectionSettings/azureWebJobsStorage/AzureWebJobsStorageExecuteStep';
-import { AzureWebJobsStoragePromptStep } from '../commands/appSettings/connectionSettings/azureWebJobsStorage/AzureWebJobsStoragePromptStep';
-import { IAzureWebJobsStorageWizardContext } from '../commands/appSettings/connectionSettings/azureWebJobsStorage/IAzureWebJobsStorageWizardContext';
-import { IValidateConnectionOptions } from '../commands/appSettings/connectionSettings/IConnectionPromptOptions';
-import { ConnectionKey, ConnectionKeyValues, localSettingsFileName } from '../constants';
+import { ConnectionKeyValues, localSettingsFileName } from '../constants';
 import { localize } from '../localize';
 import { parseJson } from '../utils/parseJson';
 
@@ -29,25 +25,6 @@ export async function getLocalConnectionString(context: IActionContext, connecti
 
     const settings: ILocalSettingsJson = await getLocalSettingsJson(context, path.join(projectPath, localSettingsFileName));
     return settings.Values && settings.Values[connectionKey];
-}
-
-// Supports validation on both 'debug' and 'deploy'
-export async function validateStorageConnection(context: IActionContext, projectPath: string, options?: IValidateConnectionOptions): Promise<void> {
-    const currentStorageConnection: string | undefined = await getLocalConnectionString(context, ConnectionKey.Storage, projectPath);
-    if (!!currentStorageConnection) {
-        if (options?.setConnectionForDeploy) {
-            Object.assign(context, { azureWebJobsRemoteConnection: currentStorageConnection });
-        }
-        return;
-    }
-
-    const wizardContext: IAzureWebJobsStorageWizardContext = Object.assign(context, { projectPath });
-    const wizard: AzureWizard<IAzureWebJobsStorageWizardContext> = new AzureWizard(wizardContext, {
-        promptSteps: [new AzureWebJobsStoragePromptStep({ preselectedConnectionType: options?.preselectedConnectionType })],
-        executeSteps: [new AzureWebJobsStorageExecuteStep(options?.setConnectionForDeploy)]
-    });
-    await wizard.prompt();
-    await wizard.execute();
 }
 
 export enum MismatchBehavior {
